+++ src/01-one/tokenizer/plugin.js
@@ -1,15 +1,12 @@
 // import unicode from './unicode.js'
 // import contractions from './contractions.js'
-import isSentence from './is-sentence.js'
 import toTerms from './methods/terms.js'
 
-
 export default {
-  mutate: (world) => {
-    // world.model.one.unicode = unicode
-    // world.model.one.contractions = contractions
-    world.methods.one.tokenize.isSentence = isSentence
-  },
+  // mutate: (world) => {
+  //   world.model.one.unicode = unicode
+  //   world.model.one.contractions = contractions
+  // },
   methods: {
     toTerms: toTerms
   }